refactor(client): move API key auto-loading to test-only code
Production `AwClient::new` no longer silently reads the server's config
directory to load API keys — that coupling between client and server config
belongs only in test infrastructure.

- `load_local_api_key`, `get_server_config_dir`, and related types moved
  from `src/lib.rs` to `tests/test.rs` (test-only)
- `toml` moved from `[dependencies]` to `[dev-dependencies]`
- Added `AwClient::new_with_api_key(host, port, name, api_key)` to both
  the async and blocking clients for explicit key injection
- Auth integration test now calls `load_local_api_key` explicitly then
  creates the client with `new_with_api_key`

Addresses ErikBjare's review feedback on PR #586.
